+++ apps/public-docsite/src/pages/Controls/SpinnerPage/SpinnerPage.doc.ts
@@ -16,12 +16,14 @@ export const SpinnerPageProps: TFabricPlatformPageProps = {
     related,
   },
   ios: {
-    overview: require('!raw-loader?esModule=false!@fluentui/public-docsite/src/pages/Controls/SpinnerPage/docs/ios/SpinnerOverview.md') as string,
+    overview:
+      require('!raw-loader?esModule=false!@fluentui/public-docsite/src/pages/Controls/SpinnerPage/docs/ios/SpinnerOverview.md') as string,
     related,
     componentUrl,
   },
   android: {
-    overview: require('!raw-loader?esModule=false!@fluentui/public-docsite/src/pages/Controls/SpinnerPage/docs/android/SpinnerOverview.md') as string,
+    overview:
+      require('!raw-loader?esModule=false!@fluentui/public-docsite/src/pages/Controls/SpinnerPage/docs/android/SpinnerOverview.md') as string,
     related,
     componentUrl,
   },