+++ packages/react-components/react-components/src/index.ts
@@ -24,6 +24,7 @@ export {
 } from '@fluentui/react-provider';
 export type {
   FluentProviderContextValues,
+  FluentProviderCustomStyleHooks,
   FluentProviderProps,
   FluentProviderSlots,
   FluentProviderState,
@@ -232,6 +233,7 @@ export {
   useCounterBadge_unstable,
   useCounterBadgeStyles_unstable,
   usePresenceBadge_unstable,
+  usePresenceBadgeStyles_unstable,
 } from '@fluentui/react-badge';
 export type {
   BadgeProps,