+++ packages/react-components/react-dialog/src/components/Dialog/Dialog.cy.tsx
@@ -327,6 +327,69 @@ describe('Dialog', () => {
       cy.get(dialogTriggerOpenSelector).realClick();
       cy.get('body').should('have.css', 'overflow', 'hidden');
     });
+
+    it('should focus trap by default', () => {
+      mount(
+        <Dialog modalType="modal">
+          <DialogTrigger disableButtonEnhancement>
+            <Button id={dialogTriggerOpenId}>Open dialog</Button>
+          </DialogTrigger>
+          <DialogSurface>
+            <DialogTitle>Dialog title</DialogTitle>
+            <DialogBody>
+              Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam exercitationem cumque repellendus eaque
+              est dolor eius expedita nulla ullam? Tenetur reprehenderit aut voluptatum impedit voluptates in natus iure
+              cumque eaque?
+            </DialogBody>
+            <DialogActions>
+              <DialogTrigger disableButtonEnhancement>
+                <Button id={dialogTriggerCloseId} appearance="secondary">
+                  Close
+                </Button>
+              </DialogTrigger>
+              <Button id="do-something-btn" appearance="primary">
+                Do Something
+              </Button>
+            </DialogActions>
+          </DialogSurface>
+        </Dialog>,
+      );
+      cy.get(dialogTriggerOpenSelector).realClick();
+      cy.get(dialogTriggerCloseSelector).should('be.focused').realPress('Tab');
+      cy.get('#do-something-btn').should('be.focused').realPress('Tab');
+      cy.get(dialogTriggerCloseSelector).should('be.focused');
+    });
+    it('should focus on window after last element when inertTrapFocus=true', () => {
+      mount(
+        <Dialog inertTrapFocus modalType="modal">
+          <DialogTrigger disableButtonEnhancement>
+            <Button id={dialogTriggerOpenId}>Open dialog</Button>
+          </DialogTrigger>
+          <DialogSurface>
+            <DialogTitle>Dialog title</DialogTitle>
+            <DialogBody>
+              Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam exercitationem cumque repellendus eaque
+              est dolor eius expedita nulla ullam? Tenetur reprehenderit aut voluptatum impedit voluptates in natus iure
+              cumque eaque?
+            </DialogBody>
+            <DialogActions>
+              <DialogTrigger disableButtonEnhancement>
+                <Button id={dialogTriggerCloseId} appearance="secondary">
+                  Close
+                </Button>
+              </DialogTrigger>
+              <Button id="do-something-btn" appearance="primary">
+                Do Something
+              </Button>
+            </DialogActions>
+          </DialogSurface>
+        </Dialog>,
+      );
+      cy.get(dialogTriggerOpenSelector).realClick();
+      cy.get(dialogTriggerCloseSelector).should('be.focused').realPress('Tab');
+      cy.get('#do-something-btn').should('be.focused').realPress('Tab');
+      cy.focused().should('not.exist');
+    });
   });
   describe('modalType = non-modal', () => {
     it('should close with escape keydown', () => {
@@ -448,6 +511,68 @@ describe('Dialog', () => {
       );
       cy.get(dialogTriggerOpenSelector).realClick();
       cy.get('body').should('have.css', 'overflow', 'hidden');
+    });
+    it('should focus trap by default', () => {
+      mount(
+        <Dialog modalType="alert">
+          <DialogTrigger disableButtonEnhancement>
+            <Button id={dialogTriggerOpenId}>Open dialog</Button>
+          </DialogTrigger>
+          <DialogSurface>
+            <DialogTitle>Dialog title</DialogTitle>
+            <DialogBody>
+              Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam exercitationem cumque repellendus eaque
+              est dolor eius expedita nulla ullam? Tenetur reprehenderit aut voluptatum impedit voluptates in natus iure
+              cumque eaque?
+            </DialogBody>
+            <DialogActions>
+              <DialogTrigger disableButtonEnhancement>
+                <Button id={dialogTriggerCloseId} appearance="secondary">
+                  Close
+                </Button>
+              </DialogTrigger>
+              <Button id="do-something-btn" appearance="primary">
+                Do Something
+              </Button>
+            </DialogActions>
+          </DialogSurface>
+        </Dialog>,
+      );
+      cy.get(dialogTriggerOpenSelector).realClick();
+      cy.get(dialogTriggerCloseSelector).should('be.focused').realPress('Tab');
+      cy.get('#do-something-btn').should('be.focused').realPress('Tab');
+      cy.get(dialogTriggerCloseSelector).should('be.focused');
+    });
+    it('should focus on window after last element when inertTrapFocus=true', () => {
+      mount(
+        <Dialog inertTrapFocus modalType="alert">
+          <DialogTrigger disableButtonEnhancement>
+            <Button id={dialogTriggerOpenId}>Open dialog</Button>
+          </DialogTrigger>
+          <DialogSurface>
+            <DialogTitle>Dialog title</DialogTitle>
+            <DialogBody>
+              Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam exercitationem cumque repellendus eaque
+              est dolor eius expedita nulla ullam? Tenetur reprehenderit aut voluptatum impedit voluptates in natus iure
+              cumque eaque?
+            </DialogBody>
+            <DialogActions>
+              <DialogTrigger disableButtonEnhancement>
+                <Button id={dialogTriggerCloseId} appearance="secondary">
+                  Close
+                </Button>
+              </DialogTrigger>
+              <Button id="do-something-btn" appearance="primary">
+                Do Something
+              </Button>
+            </DialogActions>
+          </DialogSurface>
+        </Dialog>,
+      );
+      cy.get(dialogTriggerOpenSelector).realClick();
+      cy.get(dialogTriggerCloseSelector).should('be.focused').realPress('Tab');
+      cy.get('#do-something-btn').should('be.focused').realPress('Tab');
+      cy.focused().should('not.exist');
     });
   });