+++ packages/react-components/react-migration-v8-v9/src/components/Button/ToggleButtonShim.tsx
@@ -10,19 +10,18 @@ import { shimButtonProps } from './shimButtonProps';
 /**
  * Shims v8 ToggleButton to render a v9 ToggleButton
  */
-export const ToggleButtonShim: React.ForwardRefExoticComponent<
-  IButtonProps & React.RefAttributes<HTMLButtonElement>
-> = React.forwardRef((props, _ref) => {
-  const variantProps = {
-    ...props,
-    variantClassName: props.primary ? 'ms-Button--compoundPrimary' : 'ms-Button--compound',
-  };
+export const ToggleButtonShim: React.ForwardRefExoticComponent<IButtonProps & React.RefAttributes<HTMLButtonElement>> =
+  React.forwardRef((props, _ref) => {
+    const variantProps = {
+      ...props,
+      variantClassName: props.primary ? 'ms-Button--compoundPrimary' : 'ms-Button--compound',
+    };
 
-  const shimProps: ToggleButtonProps = {
-    ...shimButtonProps(variantProps),
-    checked: props.checked,
-    defaultChecked: props.defaultChecked,
-  };
+    const shimProps: ToggleButtonProps = {
+      ...shimButtonProps(variantProps),
+      checked: props.checked,
+      defaultChecked: props.defaultChecked,
+    };
 
-  return <ToggleButton {...(props as React.RefAttributes<HTMLButtonElement>)} {...shimProps} />;
-});
+    return <ToggleButton {...(props as React.RefAttributes<HTMLButtonElement>)} {...shimProps} />;
+  });