+++ packages/react-components/react-table/src/components/DataGrid/useDataGrid.ts
@@ -1,10 +1,11 @@
 import * as React from 'react';
-import { useArrowNavigationGroup } from '@fluentui/react-tabster';
+import { useArrowNavigationGroup, useFocusFinders } from '@fluentui/react-tabster';
 import type { DataGridProps, DataGridState } from './DataGrid.types';
 import { useTable_unstable } from '../Table/useTable';
 import { useTableFeatures, useTableSort, useTableSelection, useTableColumnSizing_unstable } from '../../hooks';
 import { CELL_WIDTH } from '../TableSelectionCell';
-import { useMergedRefs } from '@fluentui/react-utilities';
+import { useEventCallback, useMergedRefs } from '@fluentui/react-utilities';
+import { End, Home } from '@fluentui/keyboard-keys';
 
 /**
  * Create the state required to render DataGrid.
@@ -59,6 +60,30 @@ export const useDataGrid_unstable = (props: DataGridProps, ref: React.Ref<HTMLEl
     }),
   ]);
 
+  const innerRef = React.useRef<HTMLDivElement>(null);
+  const { findFirstFocusable, findLastFocusable } = useFocusFinders();
+  const onKeyDown = useEventCallback((e: React.KeyboardEvent<HTMLTableElement>) => {
+    props.onKeyDown?.(e);
+    if (!innerRef.current || !e.ctrlKey || e.defaultPrevented) {
+      return;
+    }
+
+    if (e.key === Home) {
+      const firstRow = innerRef.current.querySelector('[role="row"]') as HTMLElement | null;
+      if (firstRow) {
+        findFirstFocusable(firstRow)?.focus();
+      }
+    }
+
+    if (e.key === End) {
+      const rows = innerRef.current.querySelectorAll('[role="row"]');
+      if (rows.length) {
+        const lastRow = rows.item(rows.length - 1);
+        findLastFocusable(lastRow as HTMLElement)?.focus();
+      }
+    }
+  });
+
   const baseTableState = useTable_unstable(
     {
       role: 'grid',
@@ -66,8 +91,9 @@ export const useDataGrid_unstable = (props: DataGridProps, ref: React.Ref<HTMLEl
       noNativeElements: true,
       ...(navigable && keyboardNavAttr),
       ...props,
+      onKeyDown,
     },
-    useMergedRefs(ref, tableState.tableRef),
+    useMergedRefs(ref, tableState.tableRef, innerRef),
   );
 
   return {